+++ common/config/conf_domain.go
@@ -61,6 +61,7 @@ func NewDefaultDomainConfig() DomainRepoConfig {
 				{640, 480},
 				{800, 600},
 			},
+			DynamicSizing: false,
 			Types: []string{
 				"image/jpeg",
 				"image/jpg",